refactor(domain/path): PathAlgorithm 추상화로 `jgrapht` 의존성 격리
