+++ src/main/java/wooteco/subway/domain/path/PathAlgorithm.java
@@ -1,0 +1,7 @@
+package wooteco.subway.domain.path;
+
+import wooteco.subway.domain.Station;
+
+public interface PathAlgorithm {
+    Path getPath(Station source, Station target);
+}
