+++ src/main/java/wooteco/subway/ui/service/PathService.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import java.util.Map;
 import java.util.stream.Collectors;
-import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.springframework.stereotype.Service;
 
 import wooteco.subway.dao.SectionDao;
@@ -12,9 +11,9 @@ import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.path.Age;
 import wooteco.subway.domain.path.Fare;
 import wooteco.subway.domain.path.Path;
-import wooteco.subway.domain.path.PathEdge;
+import wooteco.subway.domain.path.PathAlgorithm;
+import wooteco.subway.domain.path.ShortestPath;
 import wooteco.subway.domain.section.Section;
-import wooteco.subway.domain.path.ShortestPathFactory;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.request.PathRequest;
 import wooteco.subway.dto.response.PathResponse;
@@ -32,18 +31,22 @@ public class PathService {
 
     public PathResponse getPath(PathRequest pathRequest) {
         List<Section> sections = sectionDao.findAll();
+        PathAlgorithm pathAlgorithm = makePathAlgorithmFrom(sections);
+
+        Station sourceStation = stationDao.findById(pathRequest.getSource());
+        Station targetStation = stationDao.findById(pathRequest.getTarget());
+        Path path = pathAlgorithm.getPath(sourceStation, targetStation);
+        Fare fare = path.calculateFare(new Age(pathRequest.getAge()));
+
+        return new PathResponse(StationResponse.of(path.getStations()), path.getDistance(), fare.getValue());
+    }
+
+    private PathAlgorithm makePathAlgorithmFrom(List<Section> sections) {
         Map<Section, Fare> edges = sections.stream()
                 .collect(Collectors.toMap(
                         (section) -> section,
                         (section) -> sectionDao.findExtraFareById(section.getId())
                 ));
-        DijkstraShortestPath<Station, PathEdge> shortestPath = ShortestPathFactory.getFrom(edges);
-
-        Station sourceStation = stationDao.findById(pathRequest.getSource());
-        Station targetStation = stationDao.findById(pathRequest.getTarget());
-        Path path = Path.from(shortestPath, sourceStation, targetStation);
-        Fare fare = path.calculateFare(new Age(pathRequest.getAge()));
-
-        return new PathResponse(StationResponse.of(path.getStations()), path.getDistance(), fare.getValue());
+        return new ShortestPath(edges);
     }
 }
